Refactor LivePagePreview
